Add files via upload
Mainly renew data.py so that input file are simplified in data format.  (&GroupABD and 11/inComp)
Add oid to Wall/Door/Exit, which is automatically determined by the program.  Users can modify the name of Wall/Door/Exit, but not change oid.  
Label the door selection probabily in the figure (dataTool).